refactored charts-related functions
